+++ docker/docker_entrypoint.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+alembic --raiseerr upgrade head